refactor input types
