+++ src/phoenix/server/api/mutations/dataset_mutations.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Literal, Optional, Protocol
+from typing import Any, Dict, Literal, Optional, Protocol
 
 import strawberry
 from openinference.semconv.trace import (
@@ -7,29 +7,16 @@ from openinference.semconv.trace import (
     SpanAttributes,
 )
 from sqlalchemy import insert, select
-from strawberry import UNSET
-from strawberry.relay import GlobalID
 from strawberry.scalars import JSON
 from strawberry.types import Info
 
 from phoenix.db import models
 from phoenix.server.api.context import Context
+from phoenix.server.api.input_types.AddSpansToDatasetInput import AddSpansToDatasetInput
+from phoenix.server.api.types.AddSpansToDatasetPayload import AddSpansToDatasetPayload
 from phoenix.server.api.types.Dataset import Dataset
 from phoenix.server.api.types.node import from_global_id_with_expected_type
 from phoenix.server.api.types.Span import Span
-
-
-@strawberry.input
-class AddSpansToDatasetInput:
-    dataset_id: GlobalID
-    span_ids: List[GlobalID]
-    dataset_version_description: Optional[str] = UNSET
-    dataset_version_metadata: Optional[JSON] = UNSET
-
-
-@strawberry.type
-class AddSpansToDatasetPayload:
-    dataset: Dataset
 
 
 @strawberry.type
