+++ src/phoenix/server/api/types/AddSpansToDatasetPayload.py
@@ -1,0 +1,8 @@
+import strawberry
+
+from phoenix.server.api.types.Dataset import Dataset
+
+
+@strawberry.type
+class AddSpansToDatasetPayload:
+    dataset: Dataset
